pyproject.toml
[project]
name = "agent-mailer"
version = "0.1.0"
description = "Local multi-agent async collaboration network"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "aiosqlite>=0.21",
    "markdown>=3.10.2",
    "bcrypt>=4.0",
    "pyjwt>=2.8",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.25",
    "httpx>=0.28",
]

[project.scripts]
agent-mailer = "agent_mailer.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_mailer"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
